fix: keep claim helpers outside component module
