+++ next/app/[locale]/admin/claims/_components/initial-payment-claim-review.tsx
@@ -22,37 +22,9 @@ import type {
   InitialPaymentClaimQueueRow,
 } from "@/lib/initial-payment-claims";
 
-const DATE_FORMATTER = new Intl.DateTimeFormat("en-US", {
-  dateStyle: "medium",
-  timeStyle: "short",
-  timeZone: "UTC",
-});
-const CURRENCY_FORMATTERS = new Map([
-  [
-    "BRL",
-    new Intl.NumberFormat("pt-BR", {
-      currency: "BRL",
-      maximumFractionDigits: 2,
-      minimumFractionDigits: 2,
-      style: "currency",
-    }),
-  ],
-]);
+import { formatAmount, formatDate } from "./initial-payment-claim-formatters";
 
-export function formatAmount(amount: number, currency: string) {
-  return (
-    CURRENCY_FORMATTERS.get(currency) ?? CURRENCY_FORMATTERS.get("BRL")!
-  ).format(amount / 100);
-}
-
-export function formatDate(value: string | null | undefined) {
-  if (!value) return "—";
-
-  const date = new Date(value);
-  return Number.isNaN(date.getTime()) ? "—" : DATE_FORMATTER.format(date);
-}
-
-export function getInitials(name: string | null, handle: string | null) {
+function getInitials(name: string | null, handle: string | null) {
   const source = name?.trim() || handle?.replace(/^@/, "") || "Applicant";
   const initials = source
     .split(/\s+/)
@@ -64,7 +36,7 @@ export function getInitials(name: string | null, handle: string | null) {
   return initials || "A";
 }
 
-export function getHistory(value: unknown) {
+function getHistory(value: unknown) {
   return Array.isArray(value)
     ? value.filter(
         (entry): entry is Record<string, unknown> =>
